requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
google-generativeai==0.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
